VPN Demo

+++ tarakernel/tarakernel.h
@@ -261,7 +261,7 @@ int isMeOrMine(unsigned int nIp);
 int isSubNet(unsigned int nIp);
 void debugRoutine(void);  //Defined in module_timing_operations.c, called in tarakernel.c when receiving request for status from taralink.
 int checkFixTagging(struct _PacketInspection *pPacket, bool bForwarding, const struct nf_hook_state *state);   //module_forwarding.c
-void *memAlloc(int nSize);  //Defined in module_pointer_list.c
+void *memAlloc(size_t nSize);  //Defined in module_pointer_list.c
 _Node *getNewBefore(_Node *pPointer, int nStructSize);
 _Node *getNewAfter(_Node *pPointer, int nStructSize); //Defined in module_pointer_list.c
 _Node *getLast(_Node *pPointer);  //Defined in module_pointer_list.c
